Rename and move RepairStopTimesForEachTripOperation
